+++ enclone_visual/src/lib.rs
@@ -15,7 +15,6 @@ use nix::unistd::Pid;
 use perf_stats::*;
 use std::collections::HashMap;
 use std::convert::TryInto;
-use std::env;
 use std::fs::{remove_file, File};
 use std::io::Read;
 use std::process::{Command, Stdio};
@@ -63,41 +62,12 @@ pub mod proc1;
 pub mod proc2;
 pub mod process_messages;
 pub mod share;
+pub mod snapshot;
 pub mod style;
 pub mod summary;
 pub mod svg_to_geometry;
 pub mod testsuite;
 pub mod update_restart;
-
-// Copy window image to clipboard.  If the environment variable ENCLONE_VIS_SNAPSHOT is defined,
-// also save to that file.
-
-pub fn snapshot(start: &Option<Instant>) {
-    let mut filename = "/tmp/enclone_visual_snapshot.png".to_string();
-    let mut snapshot = false;
-    for (key, value) in env::vars() {
-        if key == "ENCLONE_VIS_SNAPSHOT" {
-            snapshot = true;
-            filename = value.to_string();
-        }
-    }
-    capture_as_file(&filename, get_window_id());
-    let mut bytes = Vec::<u8>::new();
-    {
-        let mut f = File::open(&filename).unwrap();
-        f.read_to_end(&mut bytes).unwrap();
-    }
-    if !snapshot {
-        remove_file(&filename).unwrap();
-    }
-    copy_png_bytes_to_clipboard(&bytes);
-    const MIN_SLEEP: f64 = 0.4;
-    let used = elapsed(&start.unwrap());
-    if used < MIN_SLEEP {
-        let ms = ((MIN_SLEEP - used) * 1000.0).round() as u64;
-        thread::sleep(Duration::from_millis(ms));
-    }
-}
 
 const DEJAVU_WIDTH_OVER_HEIGHT: f32 = 0.5175; // there's another different value at one point
